regex: Change some internal macro names for clarity
These long names are designed to remind the coder that they have
multiple meanings.  But move the reminder text to the end, as it
obscures the purposes.

And some have two halves for the separate meanings; change the names so
the halves are split by two underscores to visually emphasize this.