regcomp.h: deal with 64 bit aligned pointer data in regex program.
We cannot safely store 64 aligned data in a regnode structure due
to the implicit 32 bit alignment the base structure forces on the
data. If we want to do this we need to ensure that is stored in
two 32 bit aligned words.

I am not sure if we really should be storing pointers in the program
like this. In most cases where we need an SV attached to a regnode
structure we store it in the 'data' array which part of the regexp
structure, and then store an index to that item in the regnode. This
allows the use of a smaller member for the index instead.

This was identified by running "make test_reonly" under the ubsan build:

    ./Configure -d -Doptimize=-g -Dusedevel -DDEBUGGING \
    -Accflags='-fsanitize=address -fsanitize=undefined \
    -ggdb3' -Aldflags='-Wl,--no-as-needed -lasan -lubsan' \
    -Dcc=ccache\ gcc -Dld=gcc